Read about puppeteer here.
The Aim of this project is to get the knowledge of automation i.e, how can we automate the repititive Work using js Commands and node module (Puppeteer).

Puppeteer is a Node library which provides a high-level API to control Chrome or Chromium over the DevTools Protocol. Puppeteer runs headless by default, but can be configured to run full (non-headless) Chrome or Chromium.